+++ e2e/opentelemetry/go/rolldice.go
@@ -1,0 +1,73 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"io"
+	"log"
+	"math/rand"
+	"net/http"
+	"strconv"
+
+	"go.opentelemetry.io/contrib/bridges/otelslog"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/trace"
+)
+
+const name = "go.opentelemetry.io/otel/example/dice"
+
+var (
+	tracer  = otel.Tracer(name)
+	meter   = otel.Meter(name)
+	logger  = otelslog.NewLogger(name)
+	rollCnt metric.Int64Counter
+)
+
+func init() {
+	var err error
+	rollCnt, err = meter.Int64Counter("dice.rolls",
+		metric.WithDescription("The number of rolls by roll value"),
+		metric.WithUnit("{roll}"))
+	if err != nil {
+		panic(err)
+	}
+}
+
+func rolldice(w http.ResponseWriter, r *http.Request) {
+	ctx, span := tracer.Start(r.Context(), "roll")
+	defer span.End()
+
+	roll := 1 + rand.Intn(6)
+
+	var msg string
+	if player := r.PathValue("player"); player != "" {
+		msg = fmt.Sprintf("%s is rolling the dice", player)
+	} else {
+		msg = "Anonymous player is rolling the dice"
+	}
+	logger.InfoContext(ctx, msg, "result", roll)
+
+	if rand.Intn(100) > 90 {
+		span.RecordError(errors.New("example error"))
+	}
+	if rand.Intn(100) > 50 {
+		span.AddEvent(
+			"feature_flag",
+			trace.WithAttributes(
+				attribute.String("feature_flag.key", "foo"),
+				attribute.String("feature_flag.context.id", "user:foo"),
+				attribute.Bool("feature_flag.result.reason.inExperiment", rand.Intn(100) > 90),
+			))
+	}
+
+	rollValueAttr := attribute.Int("roll.value", roll)
+	span.SetAttributes(rollValueAttr)
+	rollCnt.Add(ctx, 1, metric.WithAttributes(rollValueAttr))
+
+	resp := strconv.Itoa(roll) + "\n"
+	if _, err := io.WriteString(w, resp); err != nil {
+		log.Printf("Write failed: %v\n", err)
+	}
+}